#105: refactor presentation flow
