+++ src/layouts/PresenterLayout.vue
@@ -1,33 +1,20 @@
 <script setup lang="ts">
 import SlideViewer from '../components/SlideViewer/SlideViewer.vue';
-import { usePresentation } from '../stores/presentation';
-import { computed, ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { RouteNames } from '../routes';
 import { useKeyDown, WellKnownKeys } from '../composables/useKeyDown';
+import { usePresenter } from '../composables/usePresenter';
 
 const { push } = useRouter();
-const presentation = usePresentation();
-const activeSlideIndex = ref<number>(0);
-const activeSlide = computed(() => presentation.slides[activeSlideIndex.value]);
+const { activeSlide, activeSlideIndex, goForward, goBack } = usePresenter();
 
 const end = () => {
   void push({ name: RouteNames.Main });
 };
 
-const goForward = () => {
-  if (activeSlideIndex.value >= presentation.slideCount) return;
-  activeSlideIndex.value++;
-};
-
-const goBackward = () => {
-  if (activeSlideIndex.value <= 0) return;
-  activeSlideIndex.value--;
-};
-
 useKeyDown([
   { key: WellKnownKeys.Escape, then: end },
-  { key: WellKnownKeys.ArrowLeft, then: goBackward },
+  { key: WellKnownKeys.ArrowLeft, then: goBack },
   { key: WellKnownKeys.ArrowRight, then: goForward },
 ]);
 </script>
